+++ tests/filters/metadata_filter_tests.py
@@ -2,13 +2,13 @@
 
 import unittest
 
-from mdbuild import markdown_processor as mdp
+from mdbuild.renderer import MetadataFilter
 
 
 class TestMetadataFilter(unittest.TestCase):
 
     def _run_filter(self, strip_summary_tags=False):
-        return [line for line in mdp.MetadataPlugin.filter(iter(self.input), strip_summary_tags=strip_summary_tags)]
+        return [line for line in MetadataFilter.filter(iter(self.input), strip_summary_tags=strip_summary_tags)]
 
 
 class TestMetadataFilterBasics(TestMetadataFilter):
@@ -33,13 +33,13 @@ class TestMetadataFilterBasics(TestMetadataFilter):
     def test_title_and_summary(self):
         """Title and summary are extracted properly."""
         self._run_filter()
-        self.assertEqual(mdp.MetadataPlugin.title, 'my headline')
-        self.assertEqual(mdp.MetadataPlugin.summary, 'this is my summary')
+        self.assertEqual(MetadataFilter.title, 'my headline')
+        self.assertEqual(MetadataFilter.summary, 'this is my summary')
 
         # if summary is enclosed in **, markup is removed properly
         self.input[3] = '**this is my summary**'
         res = self._run_filter()
-        self.assertEqual(mdp.MetadataPlugin.summary, 'this is my summary')
+        self.assertEqual(MetadataFilter.summary, 'this is my summary')
         self.assertEqual(res, [
             '# my headline',
             '',
@@ -52,7 +52,7 @@ class TestMetadataFilterBasics(TestMetadataFilter):
 
     def test_strip_summary_tags(self):
         res = self._run_filter(strip_summary_tags=True)
-        self.assertEqual(mdp.MetadataPlugin.summary, 'this is my summary')
+        self.assertEqual(MetadataFilter.summary, 'this is my summary')
         self.assertEqual(res, [
             '# my headline',
             '',
@@ -84,11 +84,11 @@ class TestMetadataFilterMetadata(TestMetadataFilter):
     def test_metadata_extraction(self):
         self._run_filter()
 
-        self.assertEqual(mdp.MetadataPlugin.title, 'my headline')
-        self.assertEqual(mdp.MetadataPlugin.summary, 'this is my summary')
-        print(repr(mdp.MetadataPlugin.metadata))
-        self.assertEqual(mdp.MetadataPlugin.metadata['author'], 'John Doe')
-        self.assertEqual(mdp.MetadataPlugin.metadata['menu-title'], 'a shorter title')
+        self.assertEqual(MetadataFilter.title, 'my headline')
+        self.assertEqual(MetadataFilter.summary, 'this is my summary')
+        print(repr(MetadataFilter.metadata))
+        self.assertEqual(MetadataFilter.metadata['author'], 'John Doe')
+        self.assertEqual(MetadataFilter.metadata['menu-title'], 'a shorter title')
 
     def test_metadata_removal(self):
         """Metadata is removed from file."""